requirements.txt
torch>=2.0
torchvision>=0.15
numpy>=1.24
pandas>=2.0
Pillow>=9.0
tqdm>=4.65
pyyaml>=6.0
matplotlib>=3.7
seaborn>=0.12
scikit-learn>=1.2
